POINT OF SALE HOSPITAL PHARMACY MANAGEMENT SYSTEM
I developed this project for a certain Hospital in fulfillment for the award of my Bsc. The project has these features

LOGIN PAGE: The Login allow users to submit their details for authentication. Two important data to be submitted are, email(username) and password. A user is also expected to select the account type to login.

ADD USER PAGE: The administrator will be mandated to register user details into the system. Details such as Name, Email, physical address, phone number and password (the user will change this default password at his or her first login. All these fields will be filled by the system administrator, but the user can change his profile once is logged in.

MANAGE PRESCRIPTION: This page will be managed by the doctor/clinician. The doctor will use it to view all the prescriptions that have been already made and will also use it to add new prescriptions into the system

ADD MEDICINE PAGE: This page will be used by the pharmacist to add new drugs and view the drugs that are available in the store.

VIEW PRESCRIPTION PAGE: This is another pharmacist’s platform that will be used to view the prescription made by the clinician/doctor and give the drugs to the patient. The nurse will also have the same page and will use it to provide drugs to the in-patient.

MEDICINE CATEGORY PAGE: This page will allow the doctor/pharmacist to add and view the categories of drugs that they have in the store. Categories can be tablet drugs, liquid drugs and the likes.

INVOICE/SALES PAGE: This page will be used by the cashier/accountant to create an invoice and view invoices

VIEW PAYMENT PAGE: This is also an accountant/cashier’s page where he will view all the payments that have been already made and confirmed.